Move fragment to core
