+++ core.js
@@ -97,3 +97,30 @@ sprae.use = s => {
   s.signal && use(s);
   s.compile && (compile = s.compile);
 }
+
+
+// <template> fragment holder, like persisting fragment but with minimal API surface
+export const frag = (tpl) => {
+  if (tpl._) return tpl // existing tpl
+  let content = tpl.content.cloneNode(true)
+  let attr = [...tpl.attributes]
+  let _ = document.createTextNode('') // fragment holder node
+  content.append(_)
+  let childNodes = [...content.childNodes]
+  return {
+    get parentNode() { return _.parentNode },
+    _,
+    content,
+    childNodes,
+    remove() {
+      content.append(...childNodes, _)
+    },
+    replaceWith(el) {
+      if (el === content) return
+      _.before(el)
+      this.remove()
+    },
+    attributes: attr,
+    removeAttribute(name) { attr.splice(attr.findIndex(a => a.name === name), 1) }
+  }
+}
