+++ directive/text.js
@@ -1,16 +1,13 @@
-import { directive } from "../core.js";
+import { directive, frag } from "../core.js";
 import { effect } from "../signal.js";
-import { tplfrag } from "./each.js";
 
 // set text content
 directive.text = (el, evaluate, state) => {
-  // take over tpl holder, if persistent fragment passed from each
-  if (el.holder) el = el.holder
-  // <template :text="abc"/>
-  else if (el.content) el.replaceWith(el = document.createTextNode(''))
+  // <template :text="a"/> or previously initialized template
+  if (el.content) el.replaceWith((el = frag(el)).content);
 
   return effect(() => {
     let value = evaluate(state);
-    el.textContent = value == null ? "" : value;
+    (el._ || el).textContent = value == null ? "" : value;
   });
 };
